Retain device for linux shutdown
Because the device gets dropped on the shutdown signal, the OS notifies
mullvad about it. Mullvad's internal library then removes the device
from its mapping, so later when the clear_routes gets called, it gets
confused about the device id that needs to be removed, so instead does
it based only on the IP and mask. That means that it will remove the
0.0.0.0/0 of not only the route that we added, but also of the initial
default route of the system, on Linux boxes. That makes the network
unavailable after the VPN shutdown.

This is mostly due to the fact that we use Mullvad's route manager to
setup routes for a device that is not managed by the Mullvad's library.

The solution is to reconstruct the device structure and to maintain it
until the upper layers of the function stack, so that it only gets
dropped after the cleanup has taken place.