Change tabs to spaces in Windows.mk. (#3221)
* Change tabs to spaces in `Windows.mk`.

Indented lines inside `ifeq` need to spaces instead of tabs in order to
avoid the error:

```
*** commands commence before first target.  Stop.
```

* Don't use `CPU_ARCH_LOWER` as PowerShell profile can break it.

If the user has a `Profile.ps1` set, when using PowerShell to create the
lowercase version of the CPU architecture will break.  Even calling it
with `$(shell $(SHELL) -NoProfile -Command ...)` did not solve the
issue.

* Revert "Don't use `CPU_ARCH_LOWER` as PowerShell profile can break it."

This reverts commit 51d0fd52f777edfd797c033a6536187e32aecc58.

* Manually set `$(CPU_ARCH_LOWER)`