Fix MacOS load/remove bug in store (#3711)
* Fix MacOS load/remove bug in store

This bug stems from Mullvad's code, where a previous load error of a key
in `read_all_dns` function is pushed as a `None` value in the `HashMap`
and later tried to be removed (although the `None` value can signify one
of two things, either the value was not there in the first place, or
that it was there or it should be removed).
The call of `SCDynamicStoreRemoveValue` (and in general the Apple's
`SCDynamicStore` doesn't offer much information on the possible error
reasons, but the missing key is the culprit here, since we already do a
load before the remove. Hence, this shouldn't be considered an error
since the end goal of not having the value in store is achieved.
This patch might be useful to be pushed upstream as well, even if only
for having a conversation about what's the best way to handle this.

* Update changelog